requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-multipart
python-telegram-bot
requests
httpx